// src/app/views/dashboard/Setup/staffs/models/staff.ts
export class Staff {
  id?: number;
  nombres?: string;
  apellido_pa?: string;
  apellido_ma?: string;
  dni?: string;
  correo_electronico?: string;
  celular?: string;
  sexo?: string;
  fe_nacimiento?: string;
  estado_civil_id?: string;
  departamento_id?: string;
  provincia_id?: string;
  distrito_id?: string;
  grado_instruccion_id?: string;
}